Evaluate type property to test for simpl-schema structure
Along with the launch of simple-schema 2, and the migration to
using the simpl-schema npm package, a breaking change was
introduced which prevented this package from infering field
types correctly. The change made within simple-schema supports:

https://github.com/aldeed/node-simple-schema#multiple-definitions-for-one-key

This diff doesn't fully support this feature but it prevents it
from breaking the schema when simple-schema is used in the familiar
one-def-per-field patterm I'm used to.