Map RegExp type defs to String when building schema
It's not uncommon to set a regex pattern as a type when building
a schema. This practice is supported in the docs as seen here:

https://github.com/aldeed/meteor-simple-schema#regex

This diff maps these types to Strings for the purposes of generating
the schema. This way the developer can still enjoy the validation
control of using regex with seemless graphql-bridge support!